♻️ 공통 캘린더 리팩토링
- 공통 캘린더를 유연하게 변경
